+++ server/database-adapters/oracle.js
@@ -106,6 +106,13 @@ class OracleAdapter extends BaseAdapter {
 
         return res;
     }
+
+    getSql (query) {
+        const qb = new OracleQueryBuilder();
+        const sql = qb.build(query);
+
+        return sql;
+    }
 }
 
 module.exports = { OracleAdapter };